==> 01.CBS2.spatial.Rmd <==
---
title: "SiT - Mouse Coronal Brain Section (CBS2) - 01.spatial"
author: '[Kevin Lebrigand](mailto:lebrigand@ipmc.cnrs.fr)§, Joseph Bergenstråhle §, Kim Thrane §, Annelie Mollbrink, Pascal Barbry, Rainer Waldmann and Joakim Lundeberg'
date: 'Compiled: `r format(Sys.Date(), "%B %d, %Y")`'
abstract: 'The spatial landscape of gene expression isoforms in tissue sections.'
output:
  pdf_document:
    toc: yes
    toc_depth: '3'
  html_notebook:
    code_folding: none
    theme: journal
    toc: yes
    toc_depth: 3
    toc_float: yes
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
---

<style type="text/css">

body, td {
   font-size: 15px;
}
code.r{
  font-size: 15px;
}
pre {
  font-size: 15px
}
</style>

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  cache = FALSE,
  cache.lazy = FALSE,
  tidy = TRUE
)
```

# Loading Spatial data

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=4, fig.width=6}

setwd("thisdir")
source("00.common.import.R")

CBS2 <- Load10X_Spatial(data.dir = "illumina/CBS2")
#ribo.genes <- grep(pattern = "^Rp[sl]", x = rownames(data))
#mito.genes <- grep(pattern = "^mt-", x = rownames(data))
#data <- data[-c(ribo.genes,mito.genes),]

VlnPlot(CBS2, features = c("nCount_Spatial","nFeature_Spatial"), pt.size = 0.1) + NoLegend() + ggtitle("CBS2 nCount_Spatial")
SpatialFeaturePlot(CBS2, features = c("nCount_Spatial","nFeature_Spatial"), pt.size.factor = 2) + theme(legend.position = "right") + ggtitle("CBS2 nCount_Spatial")

```

# Quality control, no filtration

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=6, fig.width=16}

mito.genes <- grep(pattern = "^mt-", x = rownames(CBS2@assays$Spatial), value = TRUE)
dropouts <- Matrix::colSums(CBS2@assays$Spatial@data == 0)/nrow(CBS2@assays$Spatial)
ribo.genes <- grep(pattern = "^Rp[sl]", x = rownames(CBS2@assays$Spatial), value = TRUE)
percent.mito <- Matrix::colSums(CBS2@assays$Spatial[mito.genes, ])/Matrix::colSums(CBS2@assays$Spatial)
percent.ribo <- Matrix::colSums(CBS2@assays$Spatial[ribo.genes, ])/Matrix::colSums(CBS2@assays$Spatial)
CBS2[['percent.mito']] <- percent.mito
CBS2[['percent.ribo']] <- percent.ribo
CBS2[['dropouts']] <- dropouts

VlnPlot(CBS2, features = c("nFeature_Spatial", "nCount_Spatial","dropouts","percent.ribo","percent.mito"), ncol=5, cols = "lightsteelblue3")

```

# SCTransform normalization and clustering

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=4, fig.width=6}

CBS2 <- SCTransform(CBS2, assay = "Spatial", verbose = FALSE)
CBS2 <- RunPCA(CBS2, assay = "SCT", verbose = FALSE)
CBS2 <- RunUMAP(CBS2, reduction = "pca", dims = 1:30)
CBS2 <- FindNeighbors(CBS2, reduction = "pca", dims = 1:30)
CBS2 <- FindClusters(CBS2, verbose = FALSE, resolution = 0.4)

DimPlot(CBS2, reduction = "umap", label = TRUE) + ggtitle("CBS2")
SpatialDimPlot(CBS2, label = TRUE, label.size = 3, pt.size.factor = 1.6) + ggtitle("CBS2")

```

# Clusters re-labelling

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=4, fig.width=6}

#new.cluster.ids <- c("Midbrain","Hippocampus","Isocortex-2","Entorhinal area","Olfactory area","Fiber tracts","CA1/CA2","Isocortex-1","Optic nerve","Hypothalamus","Thalamus","DG-sg","RSPv2","CA3sp")
#names(x = new.cluster.ids) <- levels(x = CBS2)
#CBS2 <- RenameIdents(object = CBS2, new.cluster.ids)

#my_levels <- c("Midbrain","Hippocampus","Isocortex-1","Isocortex-2","Entorhinal area","Olfactory area","RSPv2","Fiber tracts","Optic nerve","Hypothalamus","Thalamus","DG-sg","CA1/CA2","CA3sp")
# Relevel object@ident
#CBS2@active.ident <- factor(x = CBS2@active.ident, levels = my_levels)

new.cluster.ids <- c("Midbrain","Isocortex-1","Hippocampus area","Isocortex-1","Olfactory area","Fiber tracts", "CA1/CA2", "Isocortex-2", "Hippocampus area", "Hypothalamus", "Thalamus", "DG", "Retrosplenial area", "CA3")
names(x = new.cluster.ids) <- levels(x = CBS2)
CBS2 <- RenameIdents(object = CBS2, new.cluster.ids)

my_levels <- c("Midbrain","Hippocampus area","Isocortex-1","Isocortex-2","Olfactory area","Fiber tracts","Retrosplenial area","CA1/CA2","Thalamus","Hypothalamus","CA3","DG")
# Relevel object@ident
CBS2@active.ident <- factor(x = CBS2@active.ident, levels = my_levels)

SpatialDimPlot(CBS2, label = TRUE, label.size = 3, pt.size.factor = 1.6) + ggtitle("CBS2")

```

# Gene Markers Heatmap

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=10, fig.width=10 }

DefaultAssay(object = CBS2) <- "Spatial"
CBS2 <- NormalizeData(CBS2)
CBS2 <- ScaleData(CBS2)
markers <- FindAllMarkers(object = CBS2, only.pos = TRUE)
top <- markers %>% group_by(cluster) %>% top_n(5, avg_log2FC)
DoHeatmap(CBS2, features=top$gene, size=3.5)
write.table(markers, file="./output/CBS2.markers.csv", sep=",")

```

# Correlation between clusters

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=6, fig.width=8}

CBS2@meta.data$ClusterName <- CBS2@active.ident

mnmat <- c()
uniq <- unique(CBS2@active.ident)
for(i in 1:length(uniq)){
  mnmat <- cbind(mnmat, apply(as.matrix(CBS2@assays$SCT@data[, CBS2@meta.data$ClusterName==uniq[i]]), 1, mean))
}

colnames(mnmat) <- as.vector(unique(CBS2@active.ident))
ct=cor(mnmat)
pheatmap(ct)

saveRDS(CBS2, "./output/CBS2.spatial.rds")

```

# Session Info

```{r sessinf}
sessionInfo()
```
